Initial commit for resurrecting variational smoother
Added some checks and some comments questioning what is going on in the source.
Added call to mesh.prepare_for_use() in constructor because neighbors were not being computed for 3D meshes.
Updated logic in reader to identify movable boundary nodes on 2D plane surfaces in 3D meshes.
Probably need to update the connectivity portion of reader to reflect this.
minJ_BC needs to be updated to handle the nodes for 3D meshes.